# backend/internal/onboarding/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Consent struct {
	ID              uuid.UUID          `json:"id"`
	UserID          uuid.UUID          `json:"user_id"`
	DocumentKey     string             `json:"document_key"`
	DocumentVersion string             `json:"document_version"`
	Source          string             `json:"source"`
	AcceptedAt      pgtype.Timestamptz `json:"accepted_at"`
	WithdrawnAt     pgtype.Timestamptz `json:"withdrawn_at"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type IdentityRef struct {
	ID        uuid.UUID          `json:"id"`
	UserID    uuid.UUID          `json:"user_id"`
	Provider  string             `json:"provider"`
	Subject   string             `json:"subject"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type OnboardingIdempotency struct {
	UserID           uuid.UUID          `json:"user_id"`
	IdempotencyKey   string             `json:"idempotency_key"`
	RequestHash      []byte             `json:"request_hash"`
	ResponseSnapshot []byte             `json:"response_snapshot"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID        uuid.UUID          `json:"id"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type UserOnboarding struct {
	UserID                 uuid.UUID          `json:"user_id"`
	AdultAttestedAt        pgtype.Timestamptz `json:"adult_attested_at"`
	TermsVersion           *string            `json:"terms_version"`
	TermsAcceptedAt        pgtype.Timestamptz `json:"terms_accepted_at"`
	PrivacyVersion         *string            `json:"privacy_version"`
	PrivacyAcceptedAt      pgtype.Timestamptz `json:"privacy_accepted_at"`
	Timezone               *string            `json:"timezone"`
	Units                  *string            `json:"units"`
	PrimaryGoal            *string            `json:"primary_goal"`
	ExperienceLevel        *string            `json:"experience_level"`
	WeeklyAvailability     *int16             `json:"weekly_availability"`
	SessionDurationMinutes *int16             `json:"session_duration_minutes"`
	EquipmentAccess        []string           `json:"equipment_access"`
	DietPreference         *string            `json:"diet_preference"`
	SafetyAcknowledgedAt   pgtype.Timestamptz `json:"safety_acknowledged_at"`
	CurrentStep            int16              `json:"current_step"`
	State                  string             `json:"state"`
	Version                int64              `json:"version"`
	CompletedAt            pgtype.Timestamptz `json:"completed_at"`
	CreatedAt              pgtype.Timestamptz `json:"created_at"`
	UpdatedAt              pgtype.Timestamptz `json:"updated_at"`
}

type UserProfile struct {
	UserID      uuid.UUID          `json:"user_id"`
	DisplayName string             `json:"display_name"`
	Version     int64              `json:"version"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}