# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RepForge API
  version: 0.2.0-milestone1
  description: Authentication and onboarding milestone contract. It is not evidence of production readiness.
  license:
    name: Proprietary - all rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local development only
tags:
  - name: Operations
    description: Unauthenticated process health and safe build metadata.
  - name: Profile
    description: Authenticated profile and consent-summary operations.
  - name: Onboarding
    description: Authenticated, resumable adult onboarding operations.
paths:
  /healthz:
    get:
      operationId: getHealth
      tags: [Operations]
      summary: Check process liveness
      security: []
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example: { status: ok }
  /readyz:
    get:
      operationId: getReadiness
      tags: [Operations]
      summary: Check required serving dependencies
      security: []
      responses:
        "200":
          description: API can serve profile traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example: { status: ready }
        "503":
          $ref: "#/components/responses/NotReady"
  /version:
    get:
      operationId: getVersion
      tags: [Operations]
      summary: Return safe build metadata
      security: []
      responses:
        "200":
          description: Build metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Version"
  /v1/me:
    get:
      operationId: getMe
      tags: [Profile]
      summary: Read the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile and versioned consent summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
              example:
                id: 01993c86-8fc9-7a5a-9b8d-3302c7e917ec
                profile:
                  displayName: Local Athlete
                  version: 1
                  createdAt: 2026-08-28T12:00:00Z
                  updatedAt: 2026-08-28T12:00:00Z
                consents:
                  - documentKey: bootstrap_data_notice
                    documentVersion: draft-local-1
                    acceptedAt: 2026-08-28T12:00:00Z
                    withdrawnAt: null
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "404":
          $ref: "#/components/responses/ProfileNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/ProfileRequestTimeout"
    patch:
      operationId: updateMe
      tags: [Profile]
      summary: Update the authenticated user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMe"
            example: { displayName: Pooja, expectedVersion: 1 }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "404":
          $ref: "#/components/responses/ProfileNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "415":
          $ref: "#/components/responses/InvalidRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/ProfileRequestTimeout"
  /v1/onboarding:
    get:
      operationId: getOnboarding
      tags: [Onboarding]
      summary: Resume the authenticated user's onboarding
      description: Creates an empty server-side onboarding record on first access and returns saved progress.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current resumable onboarding state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingState"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "404":
          $ref: "#/components/responses/OnboardingNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/OnboardingRequestTimeout"
    patch:
      operationId: updateOnboarding
      tags: [Onboarding]
      summary: Save partial onboarding progress idempotently
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 16
            maxLength: 128
            pattern: "^[A-Za-z0-9._:-]+$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOnboarding"
      responses:
        "200":
          description: Persisted onboarding state; identical replays return the original response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingState"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "404":
          $ref: "#/components/responses/OnboardingNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "415":
          $ref: "#/components/responses/InvalidRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/OnboardingRequestTimeout"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Status:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string }
    Version:
      type: object
      additionalProperties: false
      required: [version, commit, builtAt]
      properties:
        version: { type: string, minLength: 1 }
        commit:
          type: string
          pattern: "^([0-9a-f]{40}|uncommitted)$"
        builtAt: { type: string, format: date-time }
    Me:
      type: object
      additionalProperties: false
      required: [id, profile, consents]
      properties:
        id:
          type: string
          format: uuid
        profile:
          $ref: "#/components/schemas/Profile"
        consents:
          type: array
          items:
            $ref: "#/components/schemas/Consent"
    Profile:
      type: object
      additionalProperties: false
      required: [displayName, version, createdAt, updatedAt]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        version:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Consent:
      type: object
      additionalProperties: false
      required: [documentKey, documentVersion, acceptedAt, withdrawnAt]
      properties:
        documentKey: { type: string, minLength: 1, maxLength: 80 }
        documentVersion: { type: string, minLength: 1, maxLength: 80 }
        acceptedAt: { type: string, format: date-time }
        withdrawnAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
    UpdateMe:
      type: object
      additionalProperties: false
      required: [displayName, expectedVersion]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        expectedVersion:
          type: integer
          format: int64
          minimum: 1
    OnboardingState:
      type: object
      additionalProperties: false
      required:
        - userId
        - state
        - currentStep
        - version
        - adultAttestedAt
        - termsVersion
        - termsAcceptedAt
        - privacyVersion
        - privacyAcceptedAt
        - timezone
        - units
        - primaryGoal
        - experienceLevel
        - weeklyAvailability
        - sessionDurationMinutes
        - equipmentAccess
        - dietPreference
        - safetyAcknowledgedAt
        - completedAt
        - createdAt
        - updatedAt
        - requiredTermsVersion
        - requiredPrivacyVersion
      properties:
        userId: { type: string, format: uuid }
        state: { type: string, enum: [in_progress, complete] }
        currentStep: { type: integer, minimum: 1, maximum: 6 }
        version: { type: integer, format: int64, minimum: 1 }
        adultAttestedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        termsVersion:
          oneOf: [{ type: string, minLength: 1, maxLength: 80 }, { type: "null" }]
        termsAcceptedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        privacyVersion:
          oneOf: [{ type: string, minLength: 1, maxLength: 80 }, { type: "null" }]
        privacyAcceptedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        timezone:
          oneOf: [{ type: string, minLength: 1, maxLength: 64 }, { type: "null" }]
        units:
          oneOf: [{ type: string, enum: [metric, imperial] }, { type: "null" }]
        primaryGoal:
          oneOf: [{ type: string, enum: [strength, muscle, general_fitness] }, { type: "null" }]
        experienceLevel:
          oneOf: [{ type: string, enum: [beginner, intermediate] }, { type: "null" }]
        weeklyAvailability:
          oneOf: [{ type: integer, minimum: 1, maximum: 7 }, { type: "null" }]
        sessionDurationMinutes:
          oneOf: [{ type: integer, minimum: 15, maximum: 180 }, { type: "null" }]
        equipmentAccess:
          type: array
          maxItems: 8
          uniqueItems: true
          items:
            type: string
            enum: [bodyweight, dumbbells, barbell, rack, bench, cables, machines, bands]
        dietPreference:
          oneOf:
            [{ type: string, enum: [vegetarian, eggetarian, vegan, omnivore] }, { type: "null" }]
        safetyAcknowledgedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        completedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        requiredTermsVersion: { type: string, minLength: 1, maxLength: 80 }
        requiredPrivacyVersion: { type: string, minLength: 1, maxLength: 80 }
    UpdateOnboarding:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        adultAttested: { type: boolean, const: true }
        termsVersion: { type: string, minLength: 1, maxLength: 80 }
        privacyVersion: { type: string, minLength: 1, maxLength: 80 }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        units: { type: string, enum: [metric, imperial] }
        primaryGoal: { type: string, enum: [strength, muscle, general_fitness] }
        experienceLevel: { type: string, enum: [beginner, intermediate] }
        weeklyAvailability: { type: integer, minimum: 1, maximum: 7 }
        sessionDurationMinutes: { type: integer, minimum: 15, maximum: 180 }
        equipmentAccess:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            type: string
            enum: [bodyweight, dumbbells, barbell, rack, bench, cables, machines, bands]
        dietPreference:
          oneOf:
            [{ type: string, enum: [vegetarian, eggetarian, vegan, omnivore] }, { type: "null" }]
        safetyAcknowledged: { type: boolean, const: true }
        currentStep: { type: integer, minimum: 1, maximum: 6 }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, traceId]
          properties:
            code: { type: string, pattern: "^[a-z][a-z0-9_]*$" }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            traceId: { type: string, format: uuid }
  responses:
    NotReady:
      description: Required serving dependency is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            {
              error:
                {
                  code: not_ready,
                  message: The service is not ready.,
                  traceId: 01993c86-8fc9-7a5a-9b8d-3302c7e917ec,
                },
            }
    InvalidRequest:
      description: Invalid JSON, content type, fields, or size
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthenticated:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    InsufficientScope:
      description: The access token lacks the required operation scope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    IdempotencyConflict:
      description: The Idempotency-Key was previously used for a different request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ProfileNotFound:
      description: No active profile exists for this identity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    VersionConflict:
      description: The optimistic profile version is stale
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationFailed:
      description: Profile fields failed domain validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ProfileRequestTimeout:
      description: The bounded profile storage operation timed out
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    OnboardingRequestTimeout:
      description: The bounded onboarding storage operation timed out
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    OnboardingNotFound:
      description: No active onboarding owner exists for this identity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    InternalError:
      description: Safe unexpected error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }